+++ demo/app-module.ts
@@ -5,6 +5,8 @@ import { MaterialModule } from "@angular/material";
 import { BrowserModule } from "@angular/platform-browser";
 
 import { AppComponent }  from "./app-cmp";
+import { NativeCmp } from "./native-cmp";
+import { routing } from "./app.routing";
 import { Ng2CompleterModule } from "../src";
 import { Ng2CompleterMdModule } from "../src";
 
@@ -16,9 +18,10 @@ import { Ng2CompleterMdModule } from "../src";
       MaterialModule,
       Ng2CompleterMdModule,
       FormsModule,
-      HttpModule
+      HttpModule,
+      routing
   ],
-  declarations: [ AppComponent ],
+  declarations: [ AppComponent, NativeCmp],
   bootstrap: [ AppComponent ]
 })
 export class AppModule { }
